pyproject.toml
[tool.poetry]
name = "socket-de-crypto"
version = "0.1.0"
description = ""
authors = ["Charubaiel <alexandrin.a@mail.ru>"]
readme = "README.md"
packages = [{include = "socket_de_crypto"}]

[tool.poetry.dependencies]
python = "^3.8"
python-binance = "^1.0.16"
pandas = "^1.5.1"
websocket = "^0.2.1"
dagster = "^1.0.17"
dagit = "^1.0.17"
requests = "^2.28.1"
numpy = "^1.23.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
